pyproject.toml

[tool.poetry]
name = "autobyteus-rpa-client"
version = "0.1.0"
description = "RPA Client with LLM integration"
authors = ["Your Name <your.email@example.com>"]
packages = [{include = "autobyteus_rpa_client"}]

[tool.poetry.dependencies]
python = "^3.8"
grpcio = "^1.59.0"
autobyteus-rpa-proto = {path = "../autobyteus_rpa_proto", develop = true}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
